Doc.
Try not to run tests if the user specified --prefix=/usr* or something
similar, because it would not be nice(!) to clobber their
configuration file and passwords.  Eventually I think we want to run
the installed versions, but specify different paths for everything.
(This used to be commit b18703fc8e15244dd840d1a1e9455a114e3f0408)