added samr_set_domain_info and samr_unknown_2E. We now get the full account policy window in usermanager, and the framework to store all those values. I plan to add a TDB file to store them.
oh, and found that the last value in a sam_unknown_info_12_inf struct is
an uint16 and not a uint32.

andrewb: you hardcoded the MAX_PASSWORD_AGE to 21 days. We can now turn it
to a value setable in usermanager.

        J.F.
(This used to be commit 99471d25693f6672d433b90a060378f6faad867f)