Don't pass down drep (data representation) arg from top level fn.
Decided to use tridge's PARSE_SCALARS and PARSE_BUFFERS pattern for
generating code for structures and function calls.  For functions, dissect
sub-structures with PARSE_SCALARS|PARSE_BUFFERS.  For arrays and structures
in structures (?) use PARSE_SCALARS on each element of the structure then
PARSE_BUFFERS.

Determine whether a type is a "simple" type or a struct/union.
(This used to be commit bb5bab58177fafd670dba8fe755b6c4753713efd)