Update the NT_STATUS -> DOS error table.
This new table is rather different to the old one (see diff posted to the
list for a sorted list of differences) and needs a *lot* of testing.

It does however seem to line up much better with what NT is using, as
exampled by the change to the OBJECT_NAME_COLLISION DOS error, it now matches
win2k where it didn't before.

I can't see any critical errors we now get wrong, and I know that the auth
errors are correct as per my on-the-wire observations.

This table was produced (and I hope to comment this better later) by
using the ERRMAPEXTRACT smbtorture tool, a Win2k domain member and the
'name_to_ntstatus' auth module on the HEAD PDC.  This module returned
the username as the error, and the NT box was forced to give me a dos
error becouse thats all I negotiated on that connection.  Hence the map.

Andrew Bartlett
(This used to be commit a855dfb2e0b899d03087860e5462c2aed3ca4cad)