Code too ugly to live alert!  Aooga aooga!
Removed another pointless static buffer - the number of groups a user is a
member of.  Winbindd would crash with an assert error if a (successful)
samlogon returned more than 32 groups.  Unfortunately the programmer has to
remember to free this data from the info3 after it is no longer needed.

Luckily this general problem has been fixed in HEAD/2.2 by moving to
talloc instead of malloc for unmarshalling dynamic rpc data.