These changes commited on a 'no less broken' basis.
The previous code attempted to call winbind to find out the domain sid.  This
couldn't work for a number of reasons - not the least of which was that both
the client and server ends would reject any name (in this case domain name)
without a \ in it (or lp_winbind_seperator()).

I think this is what was intended to occour.  If there is still some need to
contact winbind for this information, I suggest a new call be created for this
- as it the server-side code doesn't allow for this information to be extracted
easily in any case.

Finally, it gets in the way of the default domain code a bit - hence why I was
actually looking at it...

Andrew Bartlett
(This used to be commit 5fe1ea7f11a314a42f867a4f159c1c63c516568f)