Make smbgroupedit a little easier on the user.
We now have error messages for most of the failure cases, and the
'sid' arguments for delete and change can now take the nt group name
as a string.

Also fix up all the internal functions to be static, keeping the
namespace clean.

Andrew Bartlett
(This used to be commit f1687587010a2390b65bd5efa37cc199b67587f8)