Add -c option to specify smb.conf file (-s was already used). This will allow you to use smbpasswd if your smb.conf file is not in the compiled location.
Move all the command line agrument processing to one place.
Process args before calling lp_load services so it gets correct smb.conf file

I'm not sure of the intent of the -L option (it no longer needs to be first).
The usage message shows only root can use it, but there is no check in the
original code to see if you were root before processing the arg. I left it
so you can specify the arg as non-root but doesn't seem correct. Also
this arg is not documented in smb.conf man page.

Could people test the various forms of using smbpasswd to make sure I
didn't break anything. I tried the ones I can think of and it seems OK.