allow nmbd to register 1b name with wins server if there is an old IP there.
This causes the become_domain_master_query_success function to try to
continue to become a dmb even if we got a response to the 1b name query.
If the machine is still alive, and defending the name, we will get an
error and fail to become dmb (as we should). If there is no machine
defending, we will register the name. This was tested with broadcast as
well and I could see no problems with this. If the name query is not
tried again in broadcast mode you can get an old IP from your browse
cache and not register even though the machine is gone.