Fairly significant change to OpenPrinterEx().  The original printing code allowed a printer to have a name that was different from its sharename.  This meant that we have to iterate over all of the printer services in smb.conf and make a get_a_printer() call to get the actual printername.
The current code never allows this so I replaced the loop
to verify the printername in the open request against
lp_servicename(snum).  get_a_printer() is one of those
expensive functions that we should avoid when we can.

The main servers that will experience a speed up with this
change are those with a moderate or large number of printers
defined (possible even as few as 30!).

This fix relates to CR 362 and 414.




jerry