Use m4-mode when editinging configure.in under emacs.
Use $CC instead of $SHLD for the "test whether building a shared library
actually works".  At the moment this is failing as $SHLD (and $LDSHFLAGS -
look at the value this is set to for Solaris) are designed to be used from
a Makefile and not within configure so this test will always fail!

Tridge has suggested moving to libtool^H^H^H^H^H a shell script that
invokes the right linker with the right args and solves this problem
nicely.