Copying commit from HEAD.
My seven-year-old daughter calls me 'Captain Pedantic'.  I don't know which
is freakier... the name or the fact that a seven-year-old knows what it
means.

Small change to correct the value we place in the DGM_LENGTH field of
NBT Datagram messages.  We have been counting the full datagram, but it's
fairly clear in the RFCs that we should only count the source name,
destination name, and payload.  We've been overcharging by 14 bytes
(the size of the NBT DGM header).

This fix brings us in line with what Windows does, and what the RFCs
say should be done.  I'm a little surprised that this didn't cause any
bugs or error messages.  I guess no one actually checks this field.
(This used to be commit c06a2ece7545a9d9f8cde04745b603f7d6c6a716)