Yes, I think this should be fixed in 2.2.4 as well.  Minor bug, minor fix.
Small change to correct the value we place in the DGM_LENGTH field of
NBT Datagram messages.  We have been counting the full datagram, but it's
fairly clear in the RFCs that we should only count the source name,
destination name, and payload.  We've been overcharging by 14 bytes
(the size of the NBT DGM header).

This fix brings us in line with what Windows does, and what the RFCs
say should be done.  I'm a little surprised that this didn't cause any
bugs or error messages.  I guess no one actually checks this field.