Another localized change to AddPrinterDriver() for CR-609.
after a talk with John, we agreed that driver init data for an old version
of a driver should be deleted upon the upload of a newer version.

However, the caveat is that we also think that we should reinit
all printers bound to the driver (this would be done via a message
sent during an update_printer()).  Not doing this right now,
but will probably go ahead and add it later.