I shouldn't commit on 4 hours of sleep. Please ignore the previous accidental commit.
Fixed printer driver file upgrade decision logic. We now compare the time stamps when the
major and minor versions are the same as the previous file major and minor versions.
This situation was discovered when upgrading a Lexmark C720 PS3 driver from one version
to another.

Further research needs to be done to determine if we should just drop the version test
altogether. But, for XP this test is, even more complicated...

JF - if you have any thoughts on this, please let me know.

JohnR cr616