Had some environment problem to test the fix, so it would be really appreciated if packagers would be able to test their findsmb utility.
This fix is to make findsmb aware that the space a valid NetBIOS name character and to not get only the first part of the name as the NetBIOS name if it contains spaces.

Simo.
(This used to be commit 58cf23f4da2b36f173fcb6a695ff6798be5c589b)